update: regex components
