--- webapp/src/utils/DashboardUtil.js
@@ -1,18 +1,0 @@
-import {useState} from "react";
-import {genericRequest} from "./DataUtil";
-
-export function LoadingState () {
-    return (
-    <div className="Loading-container">
-        <div className="Spinner"></div>
-    </div>
-    );
-}
-
-export function ErrorState ({ message }) {
-    return (
-        <div className="Error-container">
-            <p>{message}</p>
-        </div>
-    );
-}
